+++ fl_ws/src/fl_slam_poc/fl_slam_poc/geometry/se3.py
@@ -1,8 +1,0 @@
-"""
-SE(3) operations - Legacy compatibility.
-
-NOTE: This module has moved to fl_slam_poc.common.transforms.se3.
-This file provides backward compatibility re-exports.
-"""
-
-from fl_slam_poc.common.transforms.se3 import *